Refactor to uv dependency management
