+++ pyproject.toml
@@ -1,21 +1,44 @@
-[build-system]
-requires = ["setuptools>=61.0"]
-build-backend = "setuptools.build_meta"
-
 [project]
-name = "src"
-version = "0.0.1"
+name = "deepweeds"
+version = "0.1.0"
 authors = [
   { name="Christopher Michael-Stokes", email="christophermichaelstokes@gmail.com" },
 ]
 description = "Python modules for object detection in keras. This is not intended for distribution."
 readme = "README.md"
-requires-python = ">=3.11"
+requires-python = "~=3.10"
+dependencies = [
+    "build>=1.2.2",
+    "gdown>=5.2.0",
+    "hydra-core>=1.3.2",
+    "jupyter>=1.1.1",
+    "keras-cv>=0.9.0",
+    "keras-tuner>=1.4.7",
+    "opencv-python>=4.10.0.84",
+    "pandas>=2.2.3",
+    "pillow>=10.4.0",
+    "plotly>=5.24.1",
+    "scikit-learn>=1.5.2",
+    "scipy>=1.14.1",
+    "tensorboard-plugin-profile>=2.17.0",
+    "tensorflow-datasets>=4.9.6",
+    "tensorflow[and-cuda]~=2.17 ; sys_platform == 'linux'",
+    "tensorflow-metal>=1.1.0 ; sys_platform == 'darwin'",
+    "tensorflow>=2.17 ; sys_platform == 'darwin'",
+    "tensorrt>=10.5.0 ; sys_platform == 'linux'",
+    "types-tensorflow>=2.17.0.20240920",
+]
 classifiers = [
     "Programming Language :: Python :: 3",
     "License :: OSI Approved :: MIT License",
     "Operating System :: OS Independent",
 ]
+
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
 
 [project.urls]
 Homepage = "https://github.com/ChristopherMichael-Stokes/deepweeds-kerascv"
